pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Daniel Santos <drsantos20@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.8"
django-rest-framework = "^0.1.0"
django-extensions = "^3.1.3"
django-debug-toolbar = "^3.2.2"
psycopg2-binary = "^2.9.2"
autopep8 = "^1.6.0"
isort = "^5.10.1"
flake8 = "^4.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
